+++ application/controllers/Account.php
@@ -124,8 +124,10 @@ class Account extends Application {
 						);
 						$this->players->add($newPlayer);
 						$this->data['pageTitle'] = "Player Registration Complete";
-						$this->session->loginMessage = "Player Registration values have passed validation!";
+						$this->session->statusMessage = "Player Registration values have passed validation!";
+
 						$this->data['staticMessage'] = "Congratulations!  Your account has been created with the username [ " . $regUsername . " ].  Simply login above to get started.";
+						$this->data['staticMessageType'] = "staticSuccess";
 					}
 				} else
 				{
@@ -192,8 +194,10 @@ class Account extends Application {
 							);
 							$this->players->update($updatePlayer);
 							$this->data['pageTitle'] = "Password Change Successful!";
-							$this->session->loginMessage = "Player password changed!";
+							$this->session->statusMessage = "Player password changed!";
+
 							$this->data['staticMessage'] = "Congratulations, " . $name . "!  Your account password has been successfully changed.  ";
+							$this->data['staticMessageType'] = "staticSuccess";
 						}
 					}
 
@@ -223,10 +227,10 @@ class Account extends Application {
 									$deleted = unlink($_SERVER['DOCUMENT_ROOT'] . $this->data['appRoot'] . "/assets/images/avatar/" . $curAvatar);
 									if ($deleted)
 									{
-										$this->session->loginMessage = "Your avatar has been successfully replaced with the uploaded avatar you've chosen.";
+										$this->session->statusMessage = "Your avatar has been successfully replaced with the uploaded avatar you've chosen.";
 									} else
 									{
-										$this->session->loginMessage = "Your avatar has been replaced, but the original file could not be deleted.  This is just a message for the administrators.";
+										$this->session->statusMessage = "Your avatar has been replaced, but the original file could not be deleted.  This is just a message for the administrators.";
 									}
 								}
 							} else
@@ -247,10 +251,10 @@ class Account extends Application {
 								$deleted = unlink($_SERVER['DOCUMENT_ROOT'] . $this->data['appRoot'] . "/assets/images/avatar/" . $curAvatar);
 								if ($deleted)
 								{
-									$this->session->loginMessage = "Your avatar has been successfully replaced with our generic avatar.";
+									$this->session->statusMessage = "Your avatar has been successfully replaced with our generic avatar.";
 								} else
 								{
-									$this->session->loginMessage = "Your avatar has been replaced, but the original file could not be deleted.  This is just a message for the administrators.";
+									$this->session->statusMessage = "Your avatar has been replaced, but the original file could not be deleted.  This is just a message for the administrators.";
 								}
 							}
 						}
@@ -264,7 +268,9 @@ class Account extends Application {
 							);
 							$this->players->update($updatePlayer);
 							$this->data['pageTitle'] = "Player Avatar Change Complete";
+
 							$this->data['staticMessage'] = "Congratulations, " . $name . "!  Your account avatar has been successfully changed.  ";
+							$this->data['staticMessageType'] = "staticSuccess";
 						}
 					}
 					// Get Player Data
@@ -278,6 +284,7 @@ class Account extends Application {
 				{
 					// not logged in
 					$this->data['staticMessage'] = "Viewing or Modifiying your account information requires you to be logged in first!";
+					$this->data['staticMessageType'] = "staticError";
 				}
 				break;
 			default: