+++ application/views/admin_agent.php
@@ -1,0 +1,22 @@
+<?php
+defined('BASEPATH') OR exit('No direct script access allowed');
+/**
+ * The Home Page
+ */
+?>
+<form method="post" >
+	<input type="submit" name="adminReturn" value="Back to Site Administration" />
+</form>
+<div id="agent_admin" class="box">
+	<h3>Agent Administration Tasks</h3>
+	To be implemented:
+	<ul>
+		<li>Force / Manual Agent Registration</li>
+		<li>Change Agent Code and/or Name</li>
+		<li>Manual Reset of all cards and peanuts</li>
+	</ul>
+
+</div>
+<form method="post" >
+	<input type="submit" name="adminReturn" value="Back to Site Administration" />
+</form>